+++ packages/pluggableWidgets/barcode-scanner-web/src/hooks/useCustomErrorMessage.ts
@@ -1,4 +1,3 @@
-import { Exception } from "@zxing/library/cjs";
 import { useState, useCallback } from "react";
 
 type ErrorCb<E extends Error = Error> = (error: E) => void;
@@ -6,22 +5,11 @@ type ErrorCb<E extends Error = Error> = (error: E) => void;
 type UseCustomErrorMessageHook = () => [string | null, ErrorCb];
 
 function getErrorMessage<E extends Error>(error: E): string | null {
-    // Error comes from `zxing`
-    if (error instanceof Exception) {
-        return "Error in barcode scanner: an unexpected error occurred while detecting a barcode in the camera media stream.";
+    if (error instanceof Error) {
+        return `Error in barcode scanner: ${error.message}`;
     }
 
-    if (error.name === "NotFoundError") {
-        return "Error in barcode scanner: no camera media devices were found.";
-    }
-
-    // Ignore error if user restricted camera access
-    if (error.name === "NotAllowedError") {
-        return null;
-    }
-
-    // Anything else will be a DOMException that we show as media stream error
-    return "Error in barcode scanner: an unexpected error occurred while retrieving the camera media stream.";
+    return "Error in barcode scanner: an unexpected error.";
 }
 
 export const useCustomErrorMessage: UseCustomErrorMessageHook = () => {