+++ packages/pluggableWidgets/barcode-scanner-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We fixed issue with camera, when it keep capturing video after widget is closed.
+
 ## [2.2.1] - 2022-07-27
 
 ### Fixed